commit motor speed changes in code

+++ Code/_sharedutils/rtwtypes.h
@@ -42,9 +42,9 @@
 typedef signed char int8_T;
 typedef unsigned char uint8_T;
 typedef short int16_T;
-typedef unsigned short uint16_T;
+typedef short unsigned int uint16_T;
 typedef int int32_T;
-typedef unsigned int uint32_T;
+typedef long unsigned int uint32_T;
 typedef float real32_T;
 typedef double real64_T;
